Separate test case for sort (#222)
* Add test_sort.py and separate test for bubble sort

* Separate test for comb_sort

* Add separate test for counting_sort

* Separate test for heap_sort

* Separate test for insertion_sort

* Separate test for merge_sort

* Separate test for quick_sort

* Separate test for selection_sort

* remove unused import

* removed unused import
