+++ sort/merge_sort.py
@@ -1,4 +1,3 @@
-import unittest
 def merge_sort(arr):
     """ Merge Sort
         Complexity: O(n log(n))
@@ -35,13 +34,3 @@ def merge(left, right):
 
    # Return result
     return arr
-
-class TestSuite(unittest.TestCase):
-    """
-        test suite for the function (above)
-    """
-    def test_merge_sort(self):
-        self.assertEqual([1, 5, 23, 57, 65, 1232],
-                         merge_sort([1, 5, 65, 23, 57, 1232]))
-if __name__ == "__main__":
-    unittest.main()
